+++ src/Veggerby.Boards/GameExtensions.cs
@@ -7,6 +7,7 @@ using Veggerby.Boards.Flows.Events;
 using Veggerby.Boards.States;
 
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Veggerby.Boards.Tests")]
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Veggerby.Boards.Benchmarks")]
 
 namespace Veggerby.Boards;